requirements.txt
alembic
bcrypt
cryptography
fastapi
psycopg2-binary
pydantic
SQLAlchemy
uvicorn
redis
httpx
pydantic-settings
pyjwt
passlib
pytest